+++ grafana-plugin/src/models/schedule/schedule.types.ts
@@ -1,3 +1,5 @@
+import dayjs from 'dayjs';
+
 import { GrafanaTeam } from 'models/grafana_team/grafana_team.types';
 import { SlackChannel } from 'models/slack_channel/slack_channel.types';
 import { User } from 'models/user/user.types';
@@ -7,6 +9,13 @@ export enum ScheduleType {
   'Calendar',
   'Ical',
   'API',
+}
+
+export interface RotationFormLiveParams {
+  rotationStart: dayjs.Dayjs;
+  shiftStart: dayjs.Dayjs;
+  shiftEnd: dayjs.Dayjs;
+  focusElementName: string;
 }
 
 export interface Schedule {